+++  examples/poke_1v1_fighter.py
@@ -21,7 +21,7 @@ import numpy as np
 from poke_env import PlayerConfiguration
 from poke_env.player import SimpleHeuristicsPlayer
 
-from p2lab.pokemon.battles import run_battles
+from p2lab.battling.battles import run_battles
 from p2lab.pokemon.premade import gen_1_pokemon
 from p2lab.pokemon.teams import generate_teams, import_pool
