+++ tests/test_battles.py
@@ -1,0 +1,15 @@
+from __future__ import annotations
+
+
+def test_battle_random_pokes():
+    assert True
+
+
+#     p = pokefactory.PokeFactory()
+#     eevee = p.make_pokemon(133, moves=["tackle", "growl"], level=5)
+#     pikachu = p.make_pokemon(25, moves=["thundershock", "growl"], level=5)
+# team1 = Team([eevee])
+# team2 = Team([pikachu])
+# #matches =
+
+# res = battle.run_battles(matches, teams, player_1, player_2, battles_per_match=10)
